Move all the reconciler-related stuff to src/reconciler
