+++ src/kubernetes_cluster/mod.rs
@@ -1,5 +1,4 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
-pub mod exec;
 pub mod proof;
 pub mod spec;
