+++ src/kubernetes_cluster/proof/cluster.rs
@@ -16,15 +16,15 @@ use crate::kubernetes_cluster::spec::{
     message::*,
     network,
     network::{multiset_contains_msg, network, NetworkState},
-    reconciler::Reconciler,
 };
-use vstd::{map::*, option::*, seq::*, set::*};
+use crate::reconciler::spec::Reconciler;
 use crate::state_machine::action::*;
 use crate::state_machine::state_machine::*;
 use crate::temporal_logic::defs::*;
 use crate::temporal_logic::rules::*;
 use builtin::*;
 use builtin_macros::*;
+use vstd::{map::*, option::*, seq::*, set::*};
 
 verus! {
