+++ src/kubernetes_cluster/proof/controller_runtime_liveness.rs
@@ -12,14 +12,14 @@ use crate::kubernetes_cluster::{
         controller::state_machine::controller,
         distributed_system::*,
         message::*,
-        reconciler::Reconciler,
     },
 };
-use vstd::{option::*, result::*};
+use crate::reconciler::spec::Reconciler;
 use crate::temporal_logic::defs::*;
 use crate::temporal_logic::rules::*;
 use builtin::*;
 use builtin_macros::*;
+use vstd::{option::*, result::*};
 
 verus! {
