+++ src/kubernetes_cluster/proof/kubernetes_api_safety.rs
@@ -2,11 +2,12 @@
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
 use crate::kubernetes_api_objects::{api_method::*, common::*, object::*};
-use crate::kubernetes_cluster::spec::{distributed_system::*, message::*, reconciler::Reconciler};
-use vstd::seq::*;
+use crate::kubernetes_cluster::spec::{distributed_system::*, message::*};
+use crate::reconciler::spec::Reconciler;
 use crate::temporal_logic::defs::*;
 use crate::temporal_logic::rules::*;
 use builtin::*;
 use builtin_macros::*;
+use vstd::seq::*;
 
 verus! {}
