+++ src/kubernetes_cluster/proof/wf1_assistant.rs
@@ -15,14 +15,14 @@ use crate::kubernetes_cluster::spec::{
     },
     kubernetes_api::state_machine::kubernetes_api,
     message::*,
-    reconciler::Reconciler,
 };
-use vstd::{option::*, seq::*, set::*};
+use crate::reconciler::spec::Reconciler;
 use crate::state_machine::action::*;
 use crate::state_machine::state_machine::*;
 use crate::temporal_logic::defs::*;
 use builtin::*;
 use builtin_macros::*;
+use vstd::{option::*, seq::*, set::*};
 
 verus! {
