+++ src/simple_controller/spec/simple_reconciler.rs
@@ -2,8 +2,9 @@
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
 use crate::kubernetes_api_objects::{api_method::*, common::*, config_map::*, object::*};
-use crate::kubernetes_cluster::spec::{message::*, reconciler::*};
+use crate::kubernetes_cluster::spec::message::*;
 use crate::pervasive_ext::string_const::*;
+use crate::reconciler::spec::*;
 use crate::state_machine::{action::*, state_machine::*};
 use crate::temporal_logic::defs::*;
 use builtin::*;
